Move validate_transaction_event_attributes into validators directory
