+++ tests/agent_features/test_asgi_w3c_trace_context.py
@@ -19,12 +19,11 @@ from newrelic.api.external_trace import ExternalTrace
 from newrelic.api.asgi_application import asgi_application
 
 from testing_support.asgi_testing import AsgiTest
-from testing_support.fixtures import (override_application_settings,
-        validate_transaction_event_attributes)
+from testing_support.fixtures import override_application_settings
 from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
 from testing_support.validators.validate_span_events import (
     validate_span_events)
-
+from testing_support.validators.validate_transaction_event_attributes import validate_transaction_event_attributes
 
 @asgi_application()
 async def target_asgi_application(scope, receive, send):
