+++ tests/agent_features/test_w3c_trace_context.py
@@ -19,11 +19,11 @@ import pytest
 from newrelic.api.transaction import current_transaction
 from newrelic.api.external_trace import ExternalTrace
 from newrelic.api.wsgi_application import wsgi_application
-from testing_support.fixtures import (override_application_settings,
-        validate_transaction_event_attributes)
+from testing_support.fixtures import override_application_settings
 from testing_support.validators.validate_span_events import (
     validate_span_events)
 from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
+from testing_support.validators.validate_transaction_event_attributes import validate_transaction_event_attributes
 
 @wsgi_application()
 def target_wsgi_application(environ, start_response):
