+++ tests/framework_aiohttp/test_server.py
@@ -18,12 +18,12 @@ import aiohttp
 from newrelic.core.config import global_settings
 
 from testing_support.fixtures import (
-        validate_transaction_event_attributes,
         count_transactions, override_generic_settings,
         override_application_settings, override_ignore_status_codes)
 from testing_support.validators.validate_code_level_metrics import validate_code_level_metrics
 from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
 from testing_support.validators.validate_transaction_errors import validate_transaction_errors
+from testing_support.validators.validate_transaction_event_attributes import validate_transaction_event_attributes
 
 version_info = tuple(int(_) for _ in aiohttp.__version__.split('.')[:2])
