+++ tests/framework_grpc/test_server.py
@@ -19,11 +19,12 @@ from conftest import create_stub_and_channel
 from _test_common import create_request, wait_for_transaction_completion
 from newrelic.core.config import global_settings
 from testing_support.fixtures import (
-        validate_transaction_event_attributes, override_application_settings,
+        override_application_settings,
         override_generic_settings, function_not_called)
 from testing_support.validators.validate_code_level_metrics import validate_code_level_metrics
 from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
 from testing_support.validators.validate_transaction_errors import validate_transaction_errors
+from testing_support.validators.validate_transaction_event_attributes import validate_transaction_event_attributes
 
 def select_python_version(py2, py3):
     return six.PY3 and py3 or py2
